+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This is intended as an overview of the major changes
 
 NEXT VERSION
 ===
-- ...
+- Fixed matching bug ".gitignore" also matched "foo.gitignore"
 
 v.1.14.0
 ===